計算部分を別ファイルに切り分けた
